Семинар по картам

+++ app/src/main/java/com/example/cardgit/Photo.java
@@ -2,26 +2,25 @@ package com.example.cardgit;
 
 
 
-import android.support.v7.widget.RecyclerView;
-
 import java.io.Serializable;
 
 
 public class Photo implements Serializable
  {
 
-    private int iconUrl;
+    private long iconUrl;
 
 
 
-    public Photo(int iconSources) {
+
+    public Photo(long iconSources) {
         this.iconUrl = iconSources;
 
 
     }
 
 
-    public int getIconSources() {
+    public long getIconSources() {
         return iconUrl;
     }
